// Frontend/libs/infrastructure/src/lib/openapi/model/companyEntity.ts
/**
 * Effektive Entwicklung mit TypeScript
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PersonEntity } from './personEntity';


export interface CompanyEntity { 
    name?: string | null;
    contact?: PersonEntity;
}