// Frontend/libs/infrastructure/src/lib/openapi/model/reservationEntity.ts
/**
 * Effektive Entwicklung mit TypeScript
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ReservationContactEntity } from './reservationContactEntity';


export interface ReservationEntity { 
    id?: number | null;
    restaurantId?: number;
    date?: string;
    contact?: ReservationContactEntity;
    personcount?: number;
}